requirements.txt
flask
flask-cors
google-generativeai
mss
pyautogui
pillow
